Add files via upload
This project contains the website widget of invoid which leads the user to verify their e-KYC. Now a days there is a cyber crime in the nation worldwide and there are many fraud activities going on. So this Website lead the user to have the proper verification by uploading their legal document like passport, Aadhar card, Driver's license,Pancard,etc.